+++ popkorn/src/main/kotlin/cc/popkorn/PopKornController.kt
@@ -16,4 +16,6 @@ interface PopKornController {
 
      fun reset()
 
+     fun purge()
+
 }