+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 1.2.0 *(2019-12-XX)*
 * New: Free `Injector` from it's internal modifier. Now can be used.
 * New: Now can use parametrized methods when injecting through `InjectableProvider`  
 * New: Add propagation strategy at `Injectable` and `InjectableProvider`
+* New: Now an Injector can be purged to free memory
 
 Version 1.1.0 *(2019-11-29)*
 -----------------------------